♻️ refactor: admin 액션 기반 파일명 정리 및 에러 스키마 통일
- admin views/services/serializers를 create/delete/list 액션 기준으로 통일
- 에러 스키마는 error_serializers로 일원화하고 error.py 제거
- 삭제 API 응답코드 201→200으로 정리 (테스트 포함)
- admin urls/serializer imports 경로 갱신
