+++ apps/exams/urls/admin.py
@@ -1,16 +1,16 @@
 from django.urls import path
 
-from apps.exams.views.admin.deployments import (
+from apps.exams.views.admin.deployments_create import (
     AdminExamDeploymentCreateAPIView,
 )
 from apps.exams.views.admin.deployments_status import (
     AdminExamDeploymentStatusAPIView,
 )
-from apps.exams.views.admin.exams import AdminExamCreateAPIView
+from apps.exams.views.admin.exams_create import AdminExamCreateAPIView
 from apps.exams.views.admin.exams_delete import AdminExamDeleteAPIView
-from apps.exams.views.admin.questions import AdminExamQuestionCreateAPIView
+from apps.exams.views.admin.questions_create import AdminExamQuestionCreateAPIView
 from apps.exams.views.admin.questions_delete import AdminExamQuestionDeleteAPIView
-from apps.exams.views.admin.submissions import AdminExamSubmissionListAPIView
+from apps.exams.views.admin.submissions_list import AdminExamSubmissionListAPIView
 from apps.exams.views.admin_exam_deployment_detail_views import (
     AdminExamDeploymentDetailAPIView,
 )
